изменила в menu

+++ Laba1Gomb/Menu.cs
@@ -108,7 +108,7 @@ namespace Laba1Gomb
                 string? firm = Console.ReadLine();
 
                 Console.Write("Введите цену: ");
-                if (!float.TryParse(Console.ReadLine(), out float price))
+                if (!decimal.TryParse(Console.ReadLine(), out decimal price))
                 {
                     Console.WriteLine("Ошибка: Некорректный формат цены");
                     return;
@@ -154,7 +154,7 @@ namespace Laba1Gomb
                 string? firm = Console.ReadLine();
 
                 Console.Write("Введите цену: ");
-                if (!float.TryParse(Console.ReadLine(), out float price))
+                if (!decimal.TryParse(Console.ReadLine(), out decimal price))
                 {
                     Console.WriteLine("Ошибка: Некорректный формат цены");
                     return;
